requirements.txt
# requirements.txt (The Final Correct Version)

pandas==2.0.3
pandas-ta
python-binance
APScheduler

# Telegram Bot
python-telegram-bot==13.15

# For reading .ini config files
configparser

# --- Locked versions to resolve conflicts ---
numpy==1.23.5
xgboost==1.7.6
scikit-learn==1.3.2

# For notebooks
notebook
matplotlib